+++ changelog.md
@@ -1,3 +1,2 @@
-### Paper
-- do not save config on shutdown
-- correctly load config on startup
+### Fabric
+- fix bug, when ServerPlayer in threadLocal is null - [**#34**](https://github.com/btwonion/magnetic/issues/34)